fix a few suggestions
